app/requirements.txt
requests
azure.data.tables
